updating linting hw
