+++ students/Justin_Jameson/lesson01/assignment/inventory_management/electric_appliances_class.py
@@ -1,16 +1,22 @@
-# Furniture class
-from inventoryClass import Inventory
+# Electric appliances class
+"""
+This files is a class for Electric appliances
+"""
+from inventory_class import Inventory
+""" does this fix it?"""
 
 
-class Furniture(Inventory):
-
+class ElectricAppliances(Inventory):
+    """
+    This class is for appliances.
+    """
     def __init__(self,
                  product_code,
                  description,
                  market_price,
                  rental_price,
-                 material,
-                 size):
+                 brand,
+                 voltage):
         # Creates common instance variables from the parent class
         Inventory.__init__(self,
                            product_code,
@@ -18,16 +24,16 @@ class Furniture(Inventory):
                            market_price,
                            rental_price)
 
-        self.material = material
-        self.size = size
+        self.brand = brand
+        self.voltage = voltage
 
-    def return_as_dictionary(self):
+    def returnasdictionary(self):
         outputDict = {}
         outputDict['product_code'] = self.product_code
         outputDict['description'] = self.description
         outputDict['market_price'] = self.market_price
         outputDict['rental_price'] = self.rental_price
-        outputDict['material'] = self.material
-        outputDict['size'] = self.size
+        outputDict['brand'] = self.brand
+        outputDict['voltage'] = self.voltage
 
         return outputDict
