+++ students/Justin_Jameson/lesson01/assignment/inventory_management/furniture_class.py
@@ -1,22 +1,16 @@
-# Electric appliances class
-"""
-This files is a class for Electric appliances
-"""
-from inventoryClass import Inventory
-""" does this fix it?"""
+# Furniture class
+from inventory_class import Inventory
 
 
-class ElectricAppliances(Inventory):
-    """
-    This class is for appliances.
-    """
+class Furniture(Inventory):
+
     def __init__(self,
                  product_code,
                  description,
                  market_price,
                  rental_price,
-                 brand,
-                 voltage):
+                 material,
+                 size):
         # Creates common instance variables from the parent class
         Inventory.__init__(self,
                            product_code,
@@ -24,16 +18,16 @@ class ElectricAppliances(Inventory):
                            market_price,
                            rental_price)
 
-        self.brand = brand
-        self.voltage = voltage
+        self.material = material
+        self.size = size
 
-    def returnasdictionary(self):
+    def return_as_dictionary(self):
         outputDict = {}
         outputDict['product_code'] = self.product_code
         outputDict['description'] = self.description
         outputDict['market_price'] = self.market_price
         outputDict['rental_price'] = self.rental_price
-        outputDict['brand'] = self.brand
-        outputDict['voltage'] = self.voltage
+        outputDict['material'] = self.material
+        outputDict['size'] = self.size
 
         return outputDict
